Use local repos API / lock+unlock with fake API

+++ webpack.config.babel.js
@@ -61,7 +61,9 @@ if (devServer) {
   }
 
   function stopServer() {
-    serverProcess.kill();
+    if (serverProcess) {
+      serverProcess.kill();
+    }
     serverProcess = null;
   }
